Get use github api to be in utils
